refactor(pdf): close the remaining audit findings, with reasons
FIXED

- Repeated heading text claimed the same page twice. Searching forward from the
  cursor is not enough on its own, because the search starts AT the cursor: a
  heading appearing twice ("Overview" under two parents) matched the first
  occurrence both times. Only a repeat of the same TEXT skips past its own
  previous hit — two different headings frequently share one page, so a blanket
  advance would be wrong.

- Unbounded nesting depth. Levels arrive over IPC and `emit` recurses once per
  level, so an absurd value would nest that deep and blow the stack. In practice
  the frontend sends h1-h6, but nothing in the wire type says so. Clamped, per
  zero-trust-at-boundaries; equal levels become siblings, which the test pins.

- Four near-identical search passes collapsed into a table. The predicates
  genuinely differ; the "search forward, return on hit" wrapper around them was
  written out four times, and a fifth pass would have been a fifth copy. Order
  is load-bearing and now says so.

- outline.rs crossed 300 lines with the added rationale, so the tree shaping
  (Node, build_tree) moved to outline_tree.rs with its tests. Same seam as
  outline_match: the pure part that can be tested with no PDF at all.
  262 / 128 / 68 lines.

CLOSED WITHOUT CHANGE, reasons recorded at the code rather than in a report

- Unlocalized error strings. The rule is right for anything a user reads, but
  the only caller logs these as a warning and lets the export succeed — nothing
  reaches the UI. Translating them would put ten bundles of work on strings that
  only land in a log file, and would dilute the real user-facing set.

- Splitting `emit` into id-allocation / linkage / destination helpers. The three
  are not independent: sibling linkage needs every id in the level before any
  dictionary is built, so the seam buys naming, not decoupling.

Also pruned the stale bookmarks.rs baseline entry, which the two-way file-size
ratchet caught. My earlier removal had silently done nothing: it ran while the
shell's working directory had drifted into src-tauri, so it opened a path that
does not exist — the third time today that drift ate an edit.

Verified: 72 pdf_export tests, clippy --all-targets -D warnings clean, rustfmt
clean, file-size gate green.
